Agregar opción de carga incremental en subida de datos de bachilleres

+++ app/Http/Controllers/Ingreso/UploadFileController.php
@@ -38,7 +38,9 @@ class UploadFileController extends Controller
 
         $path = $request->file('archivo')->store('imports');
 
-        DB::delete('delete from secundaria.data_bachillerato');
+        if ($request->get('tipoCarga') === 'nueva') {
+            DB::delete('delete from secundaria.data_bachillerato');
+        }
 
         Excel::import(new BachilleratoImport, $path);
